Change set_z() parameters to uin8_t
Saves 24 bytes of flash

+++ Firmware/mesh_bed_leveling.h
@@ -29,7 +29,7 @@ public:
     // Otherwise a correction matrix is pulled from the EEPROM if available.
     static void get_meas_xy(int ix, int iy, float &x, float &y, bool use_default);
     
-    void set_z(int ix, int iy, float z) { z_values[iy][ix] = z; }
+    void set_z(uint8_t ix, uint8_t iy, float z) { z_values[iy][ix] = z; }
     
     int select_x_index(float x) {
         int i = 1;
